==> scripts/3-Hmmr-OE-initial-clustering.R <==
# Load libraries----
library(Seurat) #v4.0.1

# Load object----
load("results/objects/obj.Rdata")

# Clustering----

# SCTransform
obj <- SCTransform(obj, verbose = T, variable.features.n = 3002)

# Take Hmmr and FIJ5 out of variable features, intentional over-expression
# will influence variable features and therefore PCA and clustering
var.features <- VariableFeatures(obj, assay = "SCT")
which(var.features=="Hmmr")
which(var.features=="FIJ5")
var.features <- var.features[var.features!= "Hmmr"]
var.features <- var.features[var.features!= "FIJ5"]
which(var.features=="Hmmr")
which(var.features=="FIJ5")
VariableFeatures(obj, assay = "SCT") <- var.features

# PCA
obj <- RunPCA(obj, npcs = 50, verbose = T)
ElbowPlot(obj, ndims = 50)

# UMAP and clustering
obj <- RunUMAP(obj,
               reduction = "pca", 
               dims = 1:18,
               verbose = T)
obj <- FindNeighbors(obj,
                     dims = 1:18,
                     verbose = T)
obj <- FindClusters(obj,
                    resolution = 0.8,
                    verbose = T)
DimPlot(obj, label = T, group.by = "timepoint")

# Normalizing and scaling RNA assay
DefaultAssay(obj) <- "RNA"
obj <- NormalizeData(obj,
                     normalization.method = "LogNormalize",
                     scale.factor = 10000,
                     verbose = T)
obj <- ScaleData(obj, features = rownames(obj), verbose = T)

# Factor genotype level----
obj@meta.data$genotype <- factor(obj@meta.data$genotype, levels = c("WT", "OE"))

# Save object----
save(obj, file = "results/objects/obj.Rdata")
